Ajout tests unitaires

+++ JobOverview/Entities/Tache.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Titre { get; set; } = string.Empty;
-        public float DureePrevue { get; set; }
+        public decimal DureePrevue { get; set; }
         public decimal DureeRestante { get; set; }
         public string CodeActivite { get; set; } = string.Empty;
         public string Personne { get; set; } = string.Empty;
